requirements.txt
streamlit
python-dotenv
langchain
langchain-openai
langchain-community
langgraph
pypdf